+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## [Unreleased](https://github.com/hlpullen/quickviewer/compare/v0.1.0...HEAD)
+## [Unreleased](https://github.com/hlpullen/quickviewer/compare/v0.2.0...HEAD)
 
-### Added
+
+## [0.2.0](https://github.com/hlpullen/quickviewer/releases/tag/v0.2.0) - 2021-04-20
+
 - DICOM support for images, dose fields, and structures
 - Extra user settings: `cmap`, `dose_cmap`
 - Ability to load a dose field via the `dose` argument without loading an image file